App_Data/jobs/triggered/webjobemailsearch/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.115.2
filelock==3.12.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.23.5
openpyxl==3.1.5
orjson==3.9.1
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
psycopg2==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.15.1
pyodbc==5.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.12
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.2
rsa==4.9
ruff==0.7.0
scipy==1.10.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.40.0
termcolor==2.3.0
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.65.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
gunicorn==20.1.0
chromedriver-autoinstaller
azure-identity>=1.15.0
azure-mgmt-web>=7.0.0
